+++ AddOns/!KalielsTracker/Modules/Help.lua
@@ -14,9 +14,12 @@ KT.Help = M
 local T = LibStub("MSA-Tutorials-1.0")
 local _DBG = function(...) if _DBG then _DBG("KT", ...) end end
 
+local HELP_TITLE = KT.TITLE.." |cffffffff"..KT.VERSION.."|r"
+local HELP_PATH = KT.MEDIA_PATH.."Help\\"
+local ICON_URL = HELP_PATH.."help_icon-url"
+local ICON_HEART = "|T"..HELP_PATH.."help_patreon:14:14:2:0:256:32:174:190:0:16|t"
+
 local db, dbChar
-local helpTitle = KT.TITLE.." |cffffffff"..KT.VERSION.."|r"
-local helpPath = KT.MEDIA_PATH.."Help\\"
 local helpName = "help"
 local helpNumPages = 13
 local supportersName = "supporters"
@@ -28,6 +31,7 @@ local cWarning2 = "|cffff4200"
 local cDots = "|cff808080"
 local offs = "\n|T:1:9|t"
 local offs2 = "\n|T:1:18|t"
+local ebSpace = "|T:16:1|t\n"
 local beta = "|cffff7fff[Beta]|r"
 local new = "|cffff7fff[NEW]|r"
 
@@ -38,9 +42,9 @@ local KTF = KT.frame
 local function AddonInfo(name)
 	local info = "\nAddon "..name
 	if C_AddOns.IsAddOnLoaded(name) then
-		info = info.." |cff00ff00is installed|r. Support you can enable/disable in Options."
+		info = info.." |cff00ff00is installed|r. Support you can enable/disable in Options.\n"..ebSpace
 	else
-		info = info.." |cffff0000is not installed|r."
+		info = info.." |cffff0000is not installed|r.\n"..ebSpace
 	end
 	return info
 end
@@ -77,15 +81,18 @@ local function SetupTutorials()
 	T.RegisterTutorial(helpName, {
 		savedvariable = KT.db.global,
 		key = "helpTutorial",
-		title = helpTitle,
-		icon = helpPath.."KT_logo",
+		title = HELP_TITLE,
+		icon = HELP_PATH.."KT_logo",
 		font = "Fonts\\FRIZQT__.TTF",
-		width = 562,
-		height = 576,
+		width = 582,
+		height = 610,
+		paddingX = 35,
+		paddingTop = 26,
+		paddingBottom = 24,
 		imageWidth = 512,
 		imageHeight = 256,
 		{	-- 1
-			image = helpPath.."help_kaliels-tracker",
+			image = HELP_PATH.."help_kaliels-tracker",
 			text = cTitle..KT.TITLE.."|r is improved default Blizzard Objective Tracker.\n\n"..
 					"Some features:\n"..
 					"- Change tracker position\n"..
@@ -101,7 +108,7 @@ local function SetupTutorials()
 			shineRight = 6,
 		},
 		{	-- 2
-			image = helpPath.."help_header-buttons",
+			image = HELP_PATH.."help_header-buttons",
 			imageHeight = 128,
 			heading = "Header buttons",
 			text = "Minimize button:                                Other buttons:\n"..
@@ -116,14 +123,14 @@ local function SetupTutorials()
 					"You can set "..cBold.."[key bind]|r for Minimize button.\n"..
 					cBold.."Right Click|r on Minimize button - Focus closest Quest.\n"..
 					cBold.."Alt + Click|r on Minimize button - Open "..KT.TITLE.." Options.",
-			paddingBottom = 14,
+			paddingBottom = 18,
 			shine = KTF.MinimizeButton,
 			shineTop = 13,
 			shineBottom = -14,
 			shineRight = 15,
 		},
 		{	-- 3
-			image = helpPath.."help_quest-title-tags",
+			image = HELP_PATH.."help_quest-title-tags",
 			imageHeight = 128,
 			heading = "Quest title tags",
 			text = "At the start of quest titles you see tags like this |cffff8000[100|r|cff00b3ffhc!|r|cffff8000]|r.\n"..
@@ -134,14 +141,14 @@ local function SetupTutorials()
 					"|cff00b3ffpvp|r "..cDots.."...|r PvP quest|T:14:133|t|cff00b3ffs|r "..cDots..".......|r Scenario quest\n"..
 					"|cff00b3ffd|r "..cDots..".......|r Dungeon quest|T:14:97|t|cff00b3ffa|r "..cDots..".......|r Account quest\n"..
 					"|cff00b3ffhc|r "..cDots..".....|r Heroic quest|T:14:113|t|cff00b3ffleg|r "..cDots.."....|r Legendary quest",
-			paddingBottom = 10,
+			paddingBottom = 16,
 			shineTop = 11,
 			shineBottom = -9,
 			shineLeft = -11,
 			shineRight = 13,
 		},
 		{	-- 4
-			image = helpPath.."help_tracker-filters",
+			image = HELP_PATH.."help_tracker-filters",
 			heading = "Tracker Filters",
 			text = "For open Filters menu "..cBold.."Click|r on the button |T"..KT.MEDIA_PATH.."UI-KT-HeaderButtons:14:14:-1:1:32:64:16:30:32:46:209:170:0|t.\n\n"..
 					"There are two types of filters:\n"..
@@ -152,7 +159,6 @@ local function SetupTutorials()
 					"|cff009bffFavorites|r - Quests or Achievements now you can mark as favorites and then filter according to them.\n\n"..
 					"For Achievements can change searched categories, it will affect the outcome of the filter.\n\n"..
 					"This menu displays other options affecting the content of the tracker (e.g. options for addon PetTracker).",
-			paddingBottom = 16,
 			shine = KTF.FilterButton,
 			shineTop = 9,
 			shineBottom = -10,
@@ -160,23 +166,22 @@ local function SetupTutorials()
 			shineRight = 11,
 		},
 		{	-- 5
-			image = helpPath.."help_quest-item-buttons",
+			image = HELP_PATH.."help_quest-item-buttons",
 			heading = "Quest Item buttons",
 			text = "Buttons are out of the tracker, because Blizzard doesn't allow to work with the action buttons in the default UI.\n\n"..
-					"|T"..helpPath.."help_quest-item-buttons_2:32:32:1:0:64:32:0:32:0:32|t "..cDots.."...|r  This tag indicates quest item in quest. The number inside is for\n"..
+					"|T"..HELP_PATH.."help_quest-item-buttons_2:32:32:1:0:64:32:0:32:0:32|t "..cDots.."...|r  This tag indicates quest item in quest. The number inside is for\n"..
 					"              identification moved quest item button.\n\n"..
-					"|T"..helpPath.."help_quest-item-buttons_2:32:32:0:3:64:32:32:64:0:32|t "..cDots.."...|r  Real quest item button is moved out of the tracker to the left/right\n"..
+					"|T"..HELP_PATH.."help_quest-item-buttons_2:32:32:0:3:64:32:32:64:0:32|t "..cDots.."...|r  Real quest item button is moved out of the tracker to the left/right\n"..
 					"              side (by selected anchor point). The number is the same as for the tag.\n\n"..
 					cWarning.."Warning:|r\n"..
 					"In some situation during combat, actions around the quest item buttons paused and carried it up after a player is out of combat.",
-			paddingBottom = 18,
 			shineTop = 3,
 			shineBottom = -2,
 			shineLeft = -4,
 			shineRight = 3,
 		},
 		{	-- 6
-			image = helpPath.."help_active-button",
+			image = HELP_PATH.."help_active-button",
 			heading = "Active Button",
 			text = "Active Button is for a better use of quest items. Displays quest item button for CLOSEST quest as Extra Action Button (like Draenor zone ability).\n\n"..
 					"Features:\n"..
@@ -197,7 +202,7 @@ local function SetupTutorials()
 			shineRight = 80,
 		},
 		{	-- 7
-			image = helpPath.."help_tracker-modules",
+			image = HELP_PATH.."help_tracker-modules",
 			heading = "Modules",
 			text = cTitle.."Order of Modules|r\n\n"..
 					"Allows to change the order of modules inside the tracker. Supports all modules including external (e.g. PetTracker).\n\n\n"..
@@ -210,19 +215,35 @@ local function SetupTutorials()
 			shineRight = 6,
 		},
 		{	-- 8
-			image = helpPath.."help_addon-masque",
+			image = HELP_PATH.."help_addon-masque",
 			heading = "Support addon Masque",
 			text = "Masque adds skinning support for Quest Item buttons. It also affects the Active Button (see prev page).\n"..
 					AddonInfo("Masque"),
+			editbox = {
+				{
+					icon = ICON_URL,
+					text = "https://www.curseforge.com/wow/addons/masque",
+					width = 510,
+					bottom = 2,
+				}
+			},
 		},
 		{	-- 9
-			image = helpPath.."help_addon-pettracker",
+			image = HELP_PATH.."help_addon-pettracker",
 			heading = "Support addon PetTracker",
-			text = "PetTracker support adjusts display of zone pet tracking inside "..KT.TITLE..". It also fix some visual bugs.\n"..
+			text = "PetTracker support adjusts display of zone pet tracking inside "..KT.TITLE..".\nIt also fix some visual bugs.\n"..
 					AddonInfo("PetTracker"),
+			editbox = {
+				{
+					icon = ICON_URL,
+					text = "https://www.curseforge.com/wow/addons/pettracker",
+					width = 510,
+					bottom = 2,
+				}
+			},
 		},
 		{	-- 10
-			image = helpPath.."help_addon-tomtom",
+			image = HELP_PATH.."help_addon-tomtom",
 			heading = "Support addon TomTom",
 			text = "TomTom support combined Blizzard's POI and TomTom's Arrow.\n\n"..
 					"|TInterface\\WorldMap\\UI-QuestPoi-NumberIcons:32:32:-2:0:256:256:128:160:96:128|t+"..
@@ -237,14 +258,21 @@ local function SetupTutorials()
 					"- The newly tracked or closest quest automatically gets a waypoint.\n"..
 					"- Waypoint of untracked or abandoned quest will be removed.\n"..
 					AddonInfo("TomTom"),
-			paddingBottom = 18,
+			editbox = {
+				{
+					icon = ICON_URL,
+					text = "https://www.curseforge.com/wow/addons/tomtom",
+					width = 510,
+					bottom = 4,
+				}
+			},
 			shineTop = 10,
 			shineBottom = -10,
 			shineLeft = -11,
 			shineRight = 11,
 		},
 		{	-- 11
-			heading = "       Hacks",
+			heading = "      Hacks",
 			text = "All hacks are enabled by default, you can disable them in "..KT.TITLE.." Options (section \"Hacks\").\n\n"..
 					cWarning.."Warning:|r Hacks may affect other addons!\n\n"..
 					cTitle.."LFG Hack|r\n\n"..
@@ -260,7 +288,7 @@ local function SetupTutorials()
 					cWarning2.."Negative impacts:|r unknown in WoW 11.2.0",
 		},
 		{	-- 12
-			image = helpPath.."help_events",
+			image = HELP_PATH.."help_events",
 			heading = "Events",
 			text = "The Events module displays active ongoing and scheduled events in the tracker. These events are normally "..
 					"available on the World Map.\n\n"..
@@ -270,7 +298,7 @@ local function SetupTutorials()
 					offs.."scheduled ones.",
 		},
 		{	-- 13
-			image = helpPath.."help_whats-new_logo",
+			image = HELP_PATH.."help_whats-new_logo",
 			imageWidth = 512,
 			imageHeight = 128,
 			imageTexCoords = { 0, 1, 0, 1 },
@@ -281,8 +309,15 @@ local function SetupTutorials()
 			heading = "     What's New",
 			headingFont = "Fonts\\MORPHEUS.ttf",
 			headingSize = 26,
-			text =
-					cTitle.."Version 7.12.0|r\n"..
+			text = "If you like "..KT.TITLE..", consider supporting it on Patreon "..ICON_HEART.."|r\n"..ebSpace.."\n"..
+
+					(cTitle.."Version 7.13.0|r\n"..
+					"- CHANGED - improve Party Sync reaction\n"..
+					"- CHANGED - Options - move Slash commands to Controls section, rename one command\n"..
+					"- FIXED - error when activating Party Sync\n"..
+					"\n")..
+
+					(cTitle.."Version 7.12.0|r\n"..
 					"- ADDED - Options - new Controls section with Keybindings and Visibility rules\n"..
 					"- ADDED - Options - keybindings for tracker control (collapse, hide, focus closest quest, use active quest item)\n"..
 					"- ADDED - Options - audio channel selection and prevent overlapping audio playback\n"..
@@ -290,17 +325,24 @@ local function SetupTutorials()
 					"- ADDED - support for WoW 11.2.0.62493\n"..
 					"- ADDED - addon support - BtWQuests 2.55.0 (Open Quest Chain option in the Quest context menu)\n"..
 					"- FIXED (scenario) - Mythic+ counter not showing automatically\n"..
-					"\n"..
+					"\n")..
 
 					cTitle.."Issue reporting|r\n"..
-					"For reporting please use "..cBold.."Tickets|r instead of Comments on CurseForge.\n\n\n\n"..
+					"For reporting please use "..cBold.."Tickets|r instead of Comments on CurseForge.\n"..ebSpace.."\n"..
 
 					cWarning.."Before reporting of error, please deactivate all other addons and make sure the bug is not caused by a collision with another addon.|r",
 			editbox = {
 				{
+					icon = ICON_URL,
+					text = "https://patreon.com/kalielstracker",
+					width = 510,
+					top = 52,
+				},
+				{
+					icon = ICON_URL,
 					text = "https://www.curseforge.com/wow/addons/kaliels-tracker/issues",
-					width = 450,
-					bottom = 42,
+					width = 510,
+					bottom = 40,
 				}
 			},
 			shine = KTF.Background,
@@ -349,18 +391,20 @@ local function SetupTutorials()
 	T.RegisterTutorial("supporters", {
 		savedvariable = KT.db.global,
 		key = "supportersTutorial",
-		title = helpTitle,
-		icon = helpPath.."KT_logo",
+		title = HELP_TITLE,
+		icon = HELP_PATH.."KT_logo",
 		font = "Fonts\\FRIZQT__.TTF",
-		width = 562,
-		height = 576,
+		width = 582,
+		height = 610,
+		paddingX = 35,
+		paddingTop = 26,
+		paddingBottom = 24,
 		{	-- 1
-			heading = "       Become a Patron",
-			text = "If you like "..KT.TITLE..", support me on |cfff34a54Patreon|r.\n\n"..
-					"Click on button  |T"..helpPath.."help_patreon:20:173:0:0:256:32:0:173:0:20|t  on CurseForge addon page.\n\n"..
-					"After 10 years of working on an addon, I started Patreon. It's created as\na compensation for the amount "..
-					"of time that addon development requires.\n\n"..
-					"                                    Many thanks to all supporters  |T"..helpPath.."help_patreon:16:16:0:0:256:32:174:190:0:16|t\n\n"..
+			heading = "      Become a Patron",
+			text = "If you like "..KT.TITLE..", consider supporting it on Patreon.\n\n\n\n"..
+					"After 10 years of working on this addon, I started Patreon. It's created as\na compensation for "..
+					"the amount of time that addon development requires.\n\n"..
+					"                                    Many thanks to all supporters "..ICON_HEART.."\n\n"..
 					cTitle.."Active Patrons|r\n"..
                     SetFormatedPatronName("Epic", "Liothen", "Emerald Dream")..
                     SetFormatedPatronName("Rare", "Ian F")..
@@ -376,7 +420,14 @@ local function SetupTutorials()
 					cTitle.."Testers|r\n"..
 					SetFormatedPlayerName("Asimeria", "Drak'thul")..
 					SetFormatedPlayerName("Torresman", "Drak'thul"),
-			paddingBottom = 18,
+			editbox = {
+				{
+					icon = ICON_URL,
+					text = "https://patreon.com/kalielstracker",
+					width = 510,
+					top = 52,
+				}
+			},
 		},
 	})
 end